Extracted DimensionEditor
