+++ src/Components/ScrivitoExtensions/DimensionsEditor/DimensionsEditor.scss
@@ -1,96 +1,6 @@
 .dimensions-editor {
   overflow: hidden;
 
-  .input_group {
-    align-items: stretch;
-    box-shadow: 0 0 3px rgba(0, 0, 0, 0.2);
-    display: flex;
-    flex-wrap: wrap;
-    margin-bottom: 10px;
-    position: relative;
-    white-space: pre-line;
-    width: 130px;
-    word-break: break-all;
-  }
-
-  .input_group input[type='number'],
-  .input_group select {
-    -webkit-appearance: textfield;
-    appearance: textfield;
-    flex: 1 1 auto;
-    min-width: 0;
-    position: relative;
-    width: 1%;
-  }
-
-  input[type='number']::-webkit-inner-spin-button,
-  input[type='number']::-webkit-outer-spin-button {
-    appearance: none;
-    margin: 0;
-  }
-
-  .input_group input {
-    font-size: 14px;
-    font-style: normal;
-    padding: 6px 8px;
-    line-height: 18px;
-    word-break: break-all;
-    background: #fff;
-    border: none;
-    border-radius: 3px 0 0 3px;
-    box-shadow: 0 0 3px rgba(0, 0, 0, 0.1);
-    color: #555;
-    display: block;
-    min-height: 30px;
-    width: 100%;
-    overflow: hidden;
-    text-overflow: ellipsis;
-
-    &:focus {
-      background: #fff;
-      outline: none;
-    }
-  }
-
-  .input_group:not([aria-readonly]):focus-within {
-    border-radius: 3px;
-    box-shadow: 0 0 0 2px var(--jr-blue-light);
-  }
-
-  .input_group .input_group_text {
-    align-items: center;
-    background: #eee;
-    border-radius: 0 3px 3px 0;
-    display: flex;
-    font-size: 14px;
-    justify-content: center;
-    line-height: 18px;
-    min-width: 40px;
-    padding: 6px 8px;
-    text-align: center;
-    white-space: nowrap;
-  }
-
-  .input_group select {
-    background: #eee;
-    border: none;
-    border-radius: 0 3px 3px 0;
-    box-shadow: 0 0 3px rgba(0, 0, 0, 0.2);
-    cursor: pointer;
-    font-size: 14px;
-    outline: none;
-    padding: 0 5px;
-    margin: 0;
-    max-width: fit-content;
-    min-width: 40px;
-    text-align: center;
-
-    &:disabled {
-      appearance: none;
-      cursor: unset;
-    }
-  }
-
   .enum_attribute {
     margin: 0;
     padding: 0 0 3px 0;
@@ -169,27 +79,6 @@
   }
 
   &.scrivito_dark {
-    .input_group input {
-      background: rgba(222, 222, 222, 0.2);
-      color: #eee;
-      font-size: 13px;
-      line-height: 18px;
-      padding: 4px 5px;
-    }
-
-    .input_group .input_group_text {
-      background: rgba(222, 222, 222, 0.3);
-      color: #eee;
-      font-size: 13px;
-      padding: 4px 5px;
-    }
-
-    .input_group select {
-      background: rgba(222, 222, 222, 0.3);
-      color: #eee;
-      font-size: 13px;
-    }
-
     .enum_attribute {
       button {
         border-left: 30px solid rgba(0, 0, 0, 0.1);
